feat(player): port desktop gamepad bridge to SDL3 (#1337)
